Grouped cloud connection dialogs into their folder
